requirements.txt
py-cpuinfo
python-cpuidpy